Move logic for calling S3 select to a shared function
Multiple lambdas will use s3 select to read manifests in s3 so move querying logic into a single shared function in t4_lambda_shared.
