+++ lambdas/pkgselect/index.py
@@ -8,7 +8,7 @@ import boto3
 import pandas as pd
 
 from t4_lambda_shared.decorator import api, validate
-from t4_lambda_shared.utils import buffer_s3response, get_default_origins, make_json_response
+from t4_lambda_shared.utils import call_s3_select, get_default_origins, make_json_response
 
 SCHEMA = {
     'type': 'object',
@@ -72,47 +72,13 @@ def get_s3_client(
     return s3_client
 
 
-def call_s3_select(
-    s3_client: str,
-    bucket: str,
-    key: str,
-    logical_key_prefix: str,
-    detail=False
-):
+def fix_quotes(key: str):
     """
-    Call S3 Select to read only the logical keys from a
-    package manifest that match the desired folder path
-    prefix
+    Convert None to emptry string and replace any single quote characters
+    with double quotes for use in SQL queries.
     """
-
-    if logical_key_prefix is None:
-        logical_key_prefix = ""
-
-    prefix_length = len(logical_key_prefix)
-    sanitized = logical_key_prefix.replace("'", "''")
-
-    if detail:
-        logical_key = sanitized
-        sql_stmt = f"SELECT s.* FROM s3object s WHERE s.logical_key = '{logical_key}' LIMIT 1"
-    else:
-        prefix = sanitized
-        sql_stmt = f"SELECT SUBSTRING(s.logical_key, {prefix_length + 1}) AS logical_key FROM s3object s"
-        if prefix:
-            sql_stmt += f" WHERE SUBSTRING(s.logical_key, 1, {prefix_length}) = '{prefix}'"
-
-    print(sql_stmt)
-    response = s3_client.select_object_content(
-        Bucket=bucket,
-        Key=key,
-        ExpressionType='SQL',
-        Expression=sql_stmt,
-        InputSerialization={
-            'JSON': {'Type': 'LINES'},
-            'CompressionType': 'NONE'
-        },
-        OutputSerialization={'JSON': {'RecordDelimiter': '\n'}}
-    )
-    return response
+    sanitized = key or ""
+    return sanitized.replace("'", "''")
 
 
 @api(cors_origins=get_default_origins())
@@ -137,15 +103,16 @@ def lambda_handler(request):
 
     # Get details of a single file in the package
     if logical_key is not None:
-        response = call_s3_select(s3_client, bucket, key, logical_key, detail=True)
-        # parse and prep response
-        response_data = json.load(buffer_s3response(response))
+        sql_stmt = f"SELECT s.* FROM s3object s WHERE s.logical_key = '{fix_quotes(logical_key)}' LIMIT 1"
+        response_data = json.load(call_s3_select(s3_client, bucket, key, sql_stmt))
     else:
         # Call s3 select to fetch only logical keys matching the
         # desired prefix (folder path)
-        response = call_s3_select(s3_client, bucket, key, prefix)
-        result = buffer_s3response(response)
-
+        prefix_length = len(prefix) if prefix is not None else 0
+        sql_stmt = f"SELECT SUBSTRING(s.logical_key, {prefix_length + 1}) AS logical_key FROM s3object s"
+        if prefix:
+            sql_stmt += f" WHERE SUBSTRING(s.logical_key, 1, {prefix_length}) = '{fix_quotes(prefix)}'"
+        result = call_s3_select(s3_client, bucket, key, sql_stmt)
         # Parse the response into a logical folder view
         df = pd.read_json(result, lines=True)
         response_data = file_list_to_folder(df)
